usb-native@5.0.12: split serialport and detector
