+++ package.json
@@ -2,17 +2,17 @@
   "name": "usb-native",
   "displayName": "serialport and usb-detection native node",
   "description": "precompile cross-platform serialport and usb-detection dll libraries and load libraries according to platform",
-  "version": "5.0.7",
+  "version": "5.0.12",
   "main": "index.js",
   "repository": "https://github.com/VSChina/serialport.node",
   "dependencies": {
+    "glob": "^7.1.6",
     "serialport": "^8.0.7",
     "usb-detection": "^4.7.0"
   },
   "devDependencies": {
     "async": "^2.1.4",
     "azure-storage": "^1.4.0",
-    "glob": "^7.1.6",
     "gulp": "^4.0.2",
     "gulp-cli": "^2.2.0",
     "linux-distro": "^4.0.1",
